requirements.txt
py-clob-client
psycopg2-binary
sqlalchemy
python-dotenv
loguru
requests
